+++ app/Http/Livewire/Settings/RolesTab.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Livewire\Settings;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 use Livewire\WithPagination;
 use Spatie\Permission\Models\Role;
@@ -9,6 +10,7 @@ use Spatie\Permission\Models\Role;
 class RolesTab extends Component
 {
     use WithPagination;
+    use AuthorizesRequests;
 
     protected $paginationTheme = 'bootstrap';
     protected $listeners = ['roles-tab-selected' => 'tabSelected'];
@@ -26,7 +28,9 @@ class RolesTab extends Component
     }
 
     public function deleteRole(Role $role)
-    {        
+    {      
+        $this->authorize('delete role');
+        
         $role->delete();
     }
 }
